+++ src/main/java/com/Denfop/integration/DE/ChaosSword.java
@@ -158,7 +158,7 @@ public class ChaosSword extends ItemSword
 	}
 
 	public double getDurabilityForDisplay(ItemStack stack) {
-		return 1.0D - getEnergyStored(stack) / getMaxEnergyStored(stack);
+		return 1.0D - (double)getEnergyStored(stack) / (double)getMaxEnergyStored(stack);
 	}
 
 	public boolean hasCustomEntity(ItemStack stack) {
